+++ src/components/Sidebar/Sidebar.js
@@ -1,8 +1,9 @@
 import React, { useContext, useState } from "react";
-import { NavLink, Outlet } from "react-router-dom";
+import { NavLink, Outlet, useNavigate } from "react-router-dom";
 import {
   ClothesIcons,
   LocationIcon,
+  MenuCloseIcons,
   NavbarMarketIcon,
   NavbarReviewIcon,
   NavbarUserIcon,
@@ -11,12 +12,12 @@ import {
 } from "../../assets/icons";
 import { dressMainData } from "../../hook/ContextTeam";
 import { useHttp } from "../../hook/useHttp";
-import { useQuery } from "@tanstack/react-query";
+import { useMutation, useQuery } from "@tanstack/react-query";
 import { RussianFlag, UzbekFlag } from "../../assets";
 import { Popover } from "antd";
-import ExitModal from "./exitModal";
 import { SellerMainData } from "../../hook/SellerUserContext";
-
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 function Sidebar() {
   const { request } = useHttp()
   const [dressInfo] = useContext(dressMainData);
@@ -61,8 +62,55 @@ function Sidebar() {
     </section>
   );
 
+  const navigate = useNavigate();
+  // -----------------------Seller Delete---------------
+  const HandleLogOutSeller = useMutation(() => {
+    return request({ url: `/logout`, method: "POST", token: true });
+  });
+  const logOutHandle = () => {
+    HandleLogOutSeller.mutate({}, {
+      onSuccess: res => {
+        if (res?.message) {
+          localStorage.clear();
+          navigate("/login-seller")
+          toast.success(`${res?.message}`, {
+            position: "top-right",
+            autoClose: 3000,
+            hideProgressBar: false,
+            closeOnClick: true,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: "light",
+          });
+          // setDressInfo({ ...dressInfo, logOutSeller: false })
+        }
+      },
+      onError: err => {
+
+      }
+    })
+  }
   return (
     <div className="w-full h-full flex ">
+      <ToastContainer
+        position="top-right"
+        autoClose={5000}
+        limit={4}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="colored"
+      />
+      <div
+        className={`fixed cursor-pointer z-[999999] inset-0 w-full h-full bg-black opacity-80 ${modalOpen ? "" : "hidden"
+          }`}
+        onClick={() => setModalOpen(false)}
+      ></div>
       <div
         className={`fixed  hidden md:block w-[300px] h-[100vh] fixed top-0 left-0  border border-lightBorderColor bg-lightBgColor
     `}
@@ -323,12 +371,46 @@ function Sidebar() {
                 Выйти
               </span>
             </button>
-            <ExitModal setModalOpen={setModalOpen} modalOpen={modalOpen} />
+            {/* <ExitModal setModalOpen={setModalOpen} modalOpen={modalOpen} /> */}
 
           </div>
         </div>
       </div >
-      <div className=" w-full md:w-[calc(100%-300px)] md:ml-[300px]">
+      <div className=" w-full md:w-[calc(100%-300px)] md:ml-[300px] ">
+        <section
+          className={`absolute max-w-[90%] md:max-w-[550px] z-[1000000] mx-auto w-full flex-col h-fit bg-white fixed py-[30px] md:py-[35px] px-[20px] md:px-[50px] rounded-t-lg rounded-b-lg md:top-[50%] duration-300 overflow-hidden left-1/2 right-1/2 translate-x-[-50%] translate-y-[-50%] hidden ${modalOpen ? "bottom-0 md:flex" : "md:hidden z-[-10]"
+            }`}
+        >
+          <button
+            onClick={() => setModalOpen(false)}
+            type="button"
+            className="absolute  right-3 top-3 w-5 h-5 ">
+            <MenuCloseIcons
+              className="w-full h-full"
+              colors={"#a1a1a1"} />
+          </button>
+          <div className="flex flex-col justify-center items-center gap-y-2 ll:gap-y-4">
+            <span className=" text-black text-lg xs:text-xl not-italic font-AeonikProMedium text-center">
+              Вы уверены?
+            </span>
+          </div>
+          <div className="w-full flex items-center justify-between mt-5 xs:mt-10 gap-x-2">
+            <button
+              onClick={() => setModalOpen(false)}
+              type="button"
+              className="w-1/2 xs:w-[45%] active:scale-95  active:opacity-70 flex items-center justify-center rounded-[12px] border border-textBlueColor text-textBlueColor bg-white h-[42px] px-4  text-center text-base not-italic font-AeonikProMedium">
+              Oтмена
+            </button>
+            <button
+              onClick={logOutHandle}
+              type="button"
+              className="w-1/2 xs:w-[45%] active:scale-95  active:opacity-70 flex items-center gap-x-2 justify-center rounded-[12px] border border-textRedColor text-white bg-[#FF4747]  h-[42px] px-4  text-center text-base not-italic font-AeonikProMedium">
+              <UserExitIcon colors={"#fff"} />{" "}
+              <span>Выйти</span>
+            </button>
+          </div>
+
+        </section >
         <Outlet />
       </div>
     </div >
